Generalize collect.collect to accept any callback.
This way, compiler wrapper scripts can be written that do something other than
write to the configured sqlite3 database. For example, they might send a
request to a service.

This change also makes collect.collect more resilient to failure. Since the
module is meant to be used in a wrapper script for the compiler, it needs not
to emit errors in situations where the compiler is not doing compilation: e.g.
when --help is passed or some other thing.